server/requirements.txt
# Photo Verification Service Dependencies
# Core image processing
Pillow>=9.5.0
piexif>=1.1.3
opencv-python>=4.8.0
numpy>=1.24.0

# AI and ML libraries (simplified)
scikit-learn>=1.3.0
imagehash>=4.3.1

# GPS and location services
geopy>=2.3.0

# HTTP requests for API calls
requests>=2.31.0

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Additional utilities
python-dotenv>=1.0.0
